include web

+++ kilde-assignment-web/src/main/java/com/senior/kilde/assignment/web/pages/tranche/TrancheCreatePageInfoTab.java
@@ -22,6 +22,7 @@ import org.apache.wicket.model.PropertyModel;
 import org.apache.wicket.validation.validator.RangeValidator;
 import org.springframework.context.ApplicationContext;
 
+import java.math.BigDecimal;
 import java.util.Map;
 
 public class TrancheCreatePageInfoTab extends ContentPanel {
@@ -187,10 +188,10 @@ public class TrancheCreatePageInfoTab extends ContentPanel {
         entity.setName(this.name_value);
         entity.setDuration(this.duration_value);
         entity.setAnnualInterest(this.annual_interest_value);
-        entity.setMinimumInvestmentAmount(this.minimum_investment_amount_value);
-        entity.setMaximumInvestmentAmount(this.maximum_investment_amount_value);
-        entity.setMaximumInvestmentAmountPerInvestor(this.maximum_investment_amount_value);
-        entity.setAmountAvailableForInvestment(this.amount_available_for_investment_value);
+        entity.setMinimumInvestmentAmount(BigDecimal.valueOf(this.minimum_investment_amount_value));
+        entity.setMaximumInvestmentAmount(BigDecimal.valueOf(this.maximum_investment_amount_value));
+        entity.setMaximumInvestmentAmountPerInvestor(BigDecimal.valueOf(this.maximum_investment_amount_value));
+        entity.setAmountAvailableForInvestment(BigDecimal.valueOf(this.amount_available_for_investment_value));
         repository.save(entity);
 
         setResponsePage(TrancheBrowsePage.class);
